Update README.md
In this repository we are taking examples of javascript programs from basic level to the intermediate level. This repository is very good for those who just start their carrier in Javascript and in Web Development. In this repository their are coding of some Javascript program and the comments tells everything what is happening in the program.